+++ gen/proto/go/parca/query/v1alpha1/query_vtproto.pb.go
@@ -1775,6 +1775,16 @@ func (m *CallgraphEdge) MarshalToSizedBufferVT(dAtA []byte) (int, error) {
 		i -= len(m.unknownFields)
 		copy(dAtA[i:], m.unknownFields)
 	}
+	if m.IsCollapsed {
+		i--
+		if m.IsCollapsed {
+			dAtA[i] = 1
+		} else {
+			dAtA[i] = 0
+		}
+		i--
+		dAtA[i] = 0x28
+	}
 	if m.Cumulative != 0 {
 		i = encodeVarint(dAtA, i, uint64(m.Cumulative))
 		i--
@@ -3146,6 +3156,9 @@ func (m *CallgraphEdge) SizeVT() (n int) {
 	}
 	if m.Cumulative != 0 {
 		n += 1 + sov(uint64(m.Cumulative))
+	}
+	if m.IsCollapsed {
+		n += 2
 	}
 	if m.unknownFields != nil {
 		n += len(m.unknownFields)
@@ -6820,6 +6833,26 @@ func (m *CallgraphEdge) UnmarshalVT(dAtA []byte) error {
 					break
 				}
 			}
+		case 5:
+			if wireType != 0 {
+				return fmt.Errorf("proto: wrong wireType = %d for field IsCollapsed", wireType)
+			}
+			var v int
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflow
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				v |= int(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			m.IsCollapsed = bool(v != 0)
 		default:
 			iNdEx = preIndex
 			skippy, err := skip(dAtA[iNdEx:])